+++ mse-20190531/alibabacloud_mse20190531/models.py
@@ -21032,6 +21032,374 @@ class ListGatewaySlbResponse(TeaModel):
         return self
 
 
+class ListInstancesRequest(TeaModel):
+    def __init__(
+        self,
+        accept_language: str = None,
+        region: str = None,
+        user_id: str = None,
+    ):
+        self.accept_language = accept_language
+        self.region = region
+        self.user_id = user_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.accept_language is not None:
+            result['AcceptLanguage'] = self.accept_language
+        if self.region is not None:
+            result['Region'] = self.region
+        if self.user_id is not None:
+            result['UserId'] = self.user_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AcceptLanguage') is not None:
+            self.accept_language = m.get('AcceptLanguage')
+        if m.get('Region') is not None:
+            self.region = m.get('Region')
+        if m.get('UserId') is not None:
+            self.user_id = m.get('UserId')
+        return self
+
+
+class ListInstancesResponseBodyData(TeaModel):
+    def __init__(
+        self,
+        app_cluster_id: str = None,
+        app_version: str = None,
+        db_password: str = None,
+        db_url: str = None,
+        db_user_name: str = None,
+        end_date: int = None,
+        environment_id: int = None,
+        gmt_create: str = None,
+        gmt_modified: str = None,
+        id: int = None,
+        image_version: str = None,
+        instance_id: str = None,
+        is_created_redis: bool = None,
+        is_created_slb: bool = None,
+        k_8s_resource_id: int = None,
+        name: str = None,
+        name_server: str = None,
+        name_server_public: str = None,
+        order_id: str = None,
+        pod_cidr: str = None,
+        primary_user: str = None,
+        region: str = None,
+        replica: int = None,
+        seata_server_unique_id: str = None,
+        security_group: str = None,
+        spec: str = None,
+        start_date: str = None,
+        status: int = None,
+        status_desc: str = None,
+        store_mode: str = None,
+        vpc: str = None,
+        vswitch: str = None,
+        zone_id: str = None,
+    ):
+        self.app_cluster_id = app_cluster_id
+        self.app_version = app_version
+        self.db_password = db_password
+        self.db_url = db_url
+        self.db_user_name = db_user_name
+        self.end_date = end_date
+        self.environment_id = environment_id
+        self.gmt_create = gmt_create
+        self.gmt_modified = gmt_modified
+        self.id = id
+        self.image_version = image_version
+        self.instance_id = instance_id
+        self.is_created_redis = is_created_redis
+        self.is_created_slb = is_created_slb
+        self.k_8s_resource_id = k_8s_resource_id
+        self.name = name
+        self.name_server = name_server
+        self.name_server_public = name_server_public
+        self.order_id = order_id
+        self.pod_cidr = pod_cidr
+        self.primary_user = primary_user
+        self.region = region
+        self.replica = replica
+        self.seata_server_unique_id = seata_server_unique_id
+        self.security_group = security_group
+        self.spec = spec
+        self.start_date = start_date
+        self.status = status
+        self.status_desc = status_desc
+        self.store_mode = store_mode
+        self.vpc = vpc
+        self.vswitch = vswitch
+        self.zone_id = zone_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.app_cluster_id is not None:
+            result['AppClusterId'] = self.app_cluster_id
+        if self.app_version is not None:
+            result['AppVersion'] = self.app_version
+        if self.db_password is not None:
+            result['DbPassword'] = self.db_password
+        if self.db_url is not None:
+            result['DbUrl'] = self.db_url
+        if self.db_user_name is not None:
+            result['DbUserName'] = self.db_user_name
+        if self.end_date is not None:
+            result['EndDate'] = self.end_date
+        if self.environment_id is not None:
+            result['EnvironmentId'] = self.environment_id
+        if self.gmt_create is not None:
+            result['GmtCreate'] = self.gmt_create
+        if self.gmt_modified is not None:
+            result['GmtModified'] = self.gmt_modified
+        if self.id is not None:
+            result['Id'] = self.id
+        if self.image_version is not None:
+            result['ImageVersion'] = self.image_version
+        if self.instance_id is not None:
+            result['InstanceId'] = self.instance_id
+        if self.is_created_redis is not None:
+            result['IsCreatedRedis'] = self.is_created_redis
+        if self.is_created_slb is not None:
+            result['IsCreatedSlb'] = self.is_created_slb
+        if self.k_8s_resource_id is not None:
+            result['K8sResourceId'] = self.k_8s_resource_id
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.name_server is not None:
+            result['NameServer'] = self.name_server
+        if self.name_server_public is not None:
+            result['NameServerPublic'] = self.name_server_public
+        if self.order_id is not None:
+            result['OrderId'] = self.order_id
+        if self.pod_cidr is not None:
+            result['PodCidr'] = self.pod_cidr
+        if self.primary_user is not None:
+            result['PrimaryUser'] = self.primary_user
+        if self.region is not None:
+            result['Region'] = self.region
+        if self.replica is not None:
+            result['Replica'] = self.replica
+        if self.seata_server_unique_id is not None:
+            result['SeataServerUniqueId'] = self.seata_server_unique_id
+        if self.security_group is not None:
+            result['SecurityGroup'] = self.security_group
+        if self.spec is not None:
+            result['Spec'] = self.spec
+        if self.start_date is not None:
+            result['StartDate'] = self.start_date
+        if self.status is not None:
+            result['Status'] = self.status
+        if self.status_desc is not None:
+            result['StatusDesc'] = self.status_desc
+        if self.store_mode is not None:
+            result['StoreMode'] = self.store_mode
+        if self.vpc is not None:
+            result['Vpc'] = self.vpc
+        if self.vswitch is not None:
+            result['Vswitch'] = self.vswitch
+        if self.zone_id is not None:
+            result['ZoneId'] = self.zone_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AppClusterId') is not None:
+            self.app_cluster_id = m.get('AppClusterId')
+        if m.get('AppVersion') is not None:
+            self.app_version = m.get('AppVersion')
+        if m.get('DbPassword') is not None:
+            self.db_password = m.get('DbPassword')
+        if m.get('DbUrl') is not None:
+            self.db_url = m.get('DbUrl')
+        if m.get('DbUserName') is not None:
+            self.db_user_name = m.get('DbUserName')
+        if m.get('EndDate') is not None:
+            self.end_date = m.get('EndDate')
+        if m.get('EnvironmentId') is not None:
+            self.environment_id = m.get('EnvironmentId')
+        if m.get('GmtCreate') is not None:
+            self.gmt_create = m.get('GmtCreate')
+        if m.get('GmtModified') is not None:
+            self.gmt_modified = m.get('GmtModified')
+        if m.get('Id') is not None:
+            self.id = m.get('Id')
+        if m.get('ImageVersion') is not None:
+            self.image_version = m.get('ImageVersion')
+        if m.get('InstanceId') is not None:
+            self.instance_id = m.get('InstanceId')
+        if m.get('IsCreatedRedis') is not None:
+            self.is_created_redis = m.get('IsCreatedRedis')
+        if m.get('IsCreatedSlb') is not None:
+            self.is_created_slb = m.get('IsCreatedSlb')
+        if m.get('K8sResourceId') is not None:
+            self.k_8s_resource_id = m.get('K8sResourceId')
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+        if m.get('NameServer') is not None:
+            self.name_server = m.get('NameServer')
+        if m.get('NameServerPublic') is not None:
+            self.name_server_public = m.get('NameServerPublic')
+        if m.get('OrderId') is not None:
+            self.order_id = m.get('OrderId')
+        if m.get('PodCidr') is not None:
+            self.pod_cidr = m.get('PodCidr')
+        if m.get('PrimaryUser') is not None:
+            self.primary_user = m.get('PrimaryUser')
+        if m.get('Region') is not None:
+            self.region = m.get('Region')
+        if m.get('Replica') is not None:
+            self.replica = m.get('Replica')
+        if m.get('SeataServerUniqueId') is not None:
+            self.seata_server_unique_id = m.get('SeataServerUniqueId')
+        if m.get('SecurityGroup') is not None:
+            self.security_group = m.get('SecurityGroup')
+        if m.get('Spec') is not None:
+            self.spec = m.get('Spec')
+        if m.get('StartDate') is not None:
+            self.start_date = m.get('StartDate')
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
+        if m.get('StatusDesc') is not None:
+            self.status_desc = m.get('StatusDesc')
+        if m.get('StoreMode') is not None:
+            self.store_mode = m.get('StoreMode')
+        if m.get('Vpc') is not None:
+            self.vpc = m.get('Vpc')
+        if m.get('Vswitch') is not None:
+            self.vswitch = m.get('Vswitch')
+        if m.get('ZoneId') is not None:
+            self.zone_id = m.get('ZoneId')
+        return self
+
+
+class ListInstancesResponseBody(TeaModel):
+    def __init__(
+        self,
+        code: int = None,
+        data: List[ListInstancesResponseBodyData] = None,
+        http_status_code: int = None,
+        message: str = None,
+        request_id: str = None,
+        success: bool = None,
+    ):
+        self.code = code
+        self.data = data
+        self.http_status_code = http_status_code
+        self.message = message
+        # Id of the request
+        self.request_id = request_id
+        self.success = success
+
+    def validate(self):
+        if self.data:
+            for k in self.data:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.code is not None:
+            result['Code'] = self.code
+        result['Data'] = []
+        if self.data is not None:
+            for k in self.data:
+                result['Data'].append(k.to_map() if k else None)
+        if self.http_status_code is not None:
+            result['HttpStatusCode'] = self.http_status_code
+        if self.message is not None:
+            result['Message'] = self.message
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Code') is not None:
+            self.code = m.get('Code')
+        self.data = []
+        if m.get('Data') is not None:
+            for k in m.get('Data'):
+                temp_model = ListInstancesResponseBodyData()
+                self.data.append(temp_model.from_map(k))
+        if m.get('HttpStatusCode') is not None:
+            self.http_status_code = m.get('HttpStatusCode')
+        if m.get('Message') is not None:
+            self.message = m.get('Message')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class ListInstancesResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: ListInstancesResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        self.validate_required(self.headers, 'headers')
+        self.validate_required(self.status_code, 'status_code')
+        self.validate_required(self.body, 'body')
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = ListInstancesResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class ListListenersByConfigRequest(TeaModel):
     def __init__(
         self,
@@ -22895,6 +23263,224 @@ class ListServiceSourceResponse(TeaModel):
             self.status_code = m.get('statusCode')
         if m.get('body') is not None:
             temp_model = ListServiceSourceResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class ListVgroupsRequest(TeaModel):
+    def __init__(
+        self,
+        accept_language: str = None,
+        region: str = None,
+        user_id: str = None,
+    ):
+        self.accept_language = accept_language
+        self.region = region
+        self.user_id = user_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.accept_language is not None:
+            result['AcceptLanguage'] = self.accept_language
+        if self.region is not None:
+            result['Region'] = self.region
+        if self.user_id is not None:
+            result['UserId'] = self.user_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AcceptLanguage') is not None:
+            self.accept_language = m.get('AcceptLanguage')
+        if m.get('Region') is not None:
+            self.region = m.get('Region')
+        if m.get('UserId') is not None:
+            self.user_id = m.get('UserId')
+        return self
+
+
+class ListVgroupsResponseBodyData(TeaModel):
+    def __init__(
+        self,
+        gmt_create: str = None,
+        gmt_modified: str = None,
+        id: int = None,
+        name: str = None,
+        name_server: str = None,
+        primary_user: str = None,
+        region: str = None,
+        seata_server_unique_id: str = None,
+    ):
+        self.gmt_create = gmt_create
+        self.gmt_modified = gmt_modified
+        self.id = id
+        self.name = name
+        self.name_server = name_server
+        self.primary_user = primary_user
+        self.region = region
+        self.seata_server_unique_id = seata_server_unique_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.gmt_create is not None:
+            result['GmtCreate'] = self.gmt_create
+        if self.gmt_modified is not None:
+            result['GmtModified'] = self.gmt_modified
+        if self.id is not None:
+            result['Id'] = self.id
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.name_server is not None:
+            result['NameServer'] = self.name_server
+        if self.primary_user is not None:
+            result['PrimaryUser'] = self.primary_user
+        if self.region is not None:
+            result['Region'] = self.region
+        if self.seata_server_unique_id is not None:
+            result['SeataServerUniqueId'] = self.seata_server_unique_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('GmtCreate') is not None:
+            self.gmt_create = m.get('GmtCreate')
+        if m.get('GmtModified') is not None:
+            self.gmt_modified = m.get('GmtModified')
+        if m.get('Id') is not None:
+            self.id = m.get('Id')
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+        if m.get('NameServer') is not None:
+            self.name_server = m.get('NameServer')
+        if m.get('PrimaryUser') is not None:
+            self.primary_user = m.get('PrimaryUser')
+        if m.get('Region') is not None:
+            self.region = m.get('Region')
+        if m.get('SeataServerUniqueId') is not None:
+            self.seata_server_unique_id = m.get('SeataServerUniqueId')
+        return self
+
+
+class ListVgroupsResponseBody(TeaModel):
+    def __init__(
+        self,
+        code: int = None,
+        data: List[ListVgroupsResponseBodyData] = None,
+        http_status_code: int = None,
+        message: str = None,
+        request_id: str = None,
+        success: bool = None,
+    ):
+        self.code = code
+        self.data = data
+        self.http_status_code = http_status_code
+        self.message = message
+        # Id of the request
+        self.request_id = request_id
+        self.success = success
+
+    def validate(self):
+        if self.data:
+            for k in self.data:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.code is not None:
+            result['Code'] = self.code
+        result['Data'] = []
+        if self.data is not None:
+            for k in self.data:
+                result['Data'].append(k.to_map() if k else None)
+        if self.http_status_code is not None:
+            result['HttpStatusCode'] = self.http_status_code
+        if self.message is not None:
+            result['Message'] = self.message
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Code') is not None:
+            self.code = m.get('Code')
+        self.data = []
+        if m.get('Data') is not None:
+            for k in m.get('Data'):
+                temp_model = ListVgroupsResponseBodyData()
+                self.data.append(temp_model.from_map(k))
+        if m.get('HttpStatusCode') is not None:
+            self.http_status_code = m.get('HttpStatusCode')
+        if m.get('Message') is not None:
+            self.message = m.get('Message')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class ListVgroupsResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: ListVgroupsResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        self.validate_required(self.headers, 'headers')
+        self.validate_required(self.status_code, 'status_code')
+        self.validate_required(self.body, 'body')
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = ListVgroupsResponseBody()
             self.body = temp_model.from_map(m['body'])
         return self